Refactor- [:feature:client-charge] Apply & Fix Detekt, Ktlint Rules
Jira Tasks - [MM-67](https://mifosforge.jira.com/browse/MM-67)
