+++ androidApp/src/test/java/org/mifos/mobile/viewModels/ClientChargeViewModelTest.kt
@@ -17,7 +17,6 @@ import org.mifos.mobile.R
 import org.mifos.mobile.models.Charge
 import org.mifos.mobile.models.Page
 import org.mifos.mobile.repositories.ClientChargeRepositoryImp
-import org.mifos.mobile.feature.client_charge.viewmodel.ClientChargeViewModel
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 import org.mifos.mobile.utils.ClientChargeUiState
 import org.mockito.Mock
@@ -45,12 +44,12 @@ class ClientChargeViewModelTest {
     @Mock
     private lateinit var clientChargeUiStateObserver: Observer<ClientChargeUiState>
 
-    private lateinit var viewModel: org.mifos.mobile.feature.client_charge.viewmodel.ClientChargeViewModel
+    private lateinit var viewModel: org.mifos.mobile.feature.charge.viewmodel.ClientChargeViewModel
 
     @Before
     fun setUp() {
         MockitoAnnotations.openMocks(this)
-        viewModel = org.mifos.mobile.feature.client_charge.viewmodel.ClientChargeViewModel(
+        viewModel = org.mifos.mobile.feature.charge.viewmodel.ClientChargeViewModel(
             clientChargeRepositoryImp
         )
     }
