+++ feature/client-charge/src/main/java/org/mifos/mobile/feature/charge/utils/ClientChargeState.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2024 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.charge.utils
+
+import org.mifos.mobile.core.datastore.model.Charge
+
+internal sealed class ClientChargeState {
+    data object Loading : ClientChargeState()
+    data class Error(val message: String?) : ClientChargeState()
+    data class Success(val charges: List<Charge>) : ClientChargeState()
+}
