+++ settings.gradle.kts
@@ -41,7 +41,7 @@ include(":feature:qr")
 include(":feature:transfer-process")
 include(":feature:account")
 include(":feature:recent_transaction")
-include(":feature:client_charge")
+include(":feature:client-charge")
 include(":feature:third-party-transfer")
 include(":feature:help")
 include(":feature:notification")
